+++ package.json
@@ -20,6 +20,11 @@
     "@next-auth/prisma-adapter": "^1.0.5",
     "@prisma/client": "^4.5.0",
     "@tabler/icons": "^1.108.0",
+    "@tanstack/react-query": "^4.13.5",
+    "@trpc/client": "10.0.0-proxy-beta.26",
+    "@trpc/next": "10.0.0-proxy-beta.26",
+    "@trpc/react-query": "10.0.0-proxy-beta.26",
+    "@trpc/server": "10.0.0-proxy-beta.26",
     "argon2": "^0.30.1",
     "cookies": "^0.8.0",
     "next": "13.0.0",
