+++ src/lib/types/environment.d.ts
