+++ src/pages/api/signup.ts
@@ -1,9 +1,14 @@
 import { NextApiResponse, NextApiRequest } from 'next';
-import { prisma } from '@/lib/utils/db/prisma';
+import { prisma } from '@/utils/db/prisma';
+import { randomUUID } from 'crypto';
 import argon2 from 'argon2';
 import * as z from 'zod';
 
 const credentials = z.object({
+  name: z
+    .string()
+    .min(1, { message: 'Name cannot be empty' })
+    .max(50, { message: 'Name should be less than 50 characters' }),
   username: z
     .string()
     .min(1, { message: 'Username cannot be empty' })
@@ -26,9 +31,9 @@ type ErrorResponse = {
   message: string;
 };
 
-type LogInResponse = CredentialsResponse | ErrorResponse;
+type SignUpResponse = CredentialsResponse | ErrorResponse;
 
-export default async function handler(req: NextApiRequest, res: NextApiResponse<LogInResponse>) {
+export default async function handler(req: NextApiRequest, res: NextApiResponse<SignUpResponse>) {
   if (req.method !== 'POST') {
     return res.status(200).json({ message: 'method not allowed' });
   }
@@ -40,6 +45,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
     return res.status(400).json({ message: 'Something wrong with your input' });
   }
 
+  const name = parse.data.name;
   const username = parse.data.username;
   const password = parse.data.password;
 
@@ -50,21 +56,38 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
       }
     });
 
-    if (user && user.password) {
-      if (await argon2.verify(user.password, password)) {
-        if (user.name && user.username) {
-          return res.status(200).json({
-            name: user.name,
-            username: user.username,
-            id: user.id
-          });
-        }
-      } else {
-        return res.status(409).json({ message: 'Invalid Credentials' });
-      }
+    if (user) {
+      return res.status(409).json({ message: 'This username already exists' });
     }
 
-    return res.status(500).json({ message: 'No Such User' });
+    const hash = await argon2.hash(password);
+
+    const newUser = await prisma.user.create({
+      data: {
+        name: name,
+        username: username,
+        password: hash
+      }
+    });
+
+    await prisma.account.create({
+      data: {
+        userId: newUser.id,
+        provider: 'credentials',
+        providerAccountId: randomUUID(),
+        type: 'credentials'
+      }
+    });
+
+    if (newUser.name && newUser.username) {
+      return res.status(200).json({
+        id: newUser.id,
+        name: newUser.name,
+        username: newUser.username
+      });
+    }
+
+    return res.status(500).json({ message: 'Something went wrong' });
   } catch (error) {
     return res.status(500).json({ message: 'An Error Occured' });
   }
