+++ src/pages/api/trpc/[trpc].ts
@@ -1,0 +1,42 @@
+import { createContext } from '@/server/context';
+import * as trpcNext from '@trpc/server/adapters/next';
+import { appRouter } from '../../../server/routers/_app';
+
+// export API handler
+// export default trpcNext.createNextApiHandler({
+//   router: appRouter,
+//   createContext: () => ({})
+// });
+
+/**
+ * This file contains tRPC's HTTP response handler
+ */
+
+export default trpcNext.createNextApiHandler({
+  router: appRouter,
+  /**
+   * @link https://trpc.io/docs/context
+   */
+  createContext,
+  /**
+   * @link https://trpc.io/docs/error-handling
+   */
+  onError({ error }) {
+    if (error.code === 'INTERNAL_SERVER_ERROR') {
+      // send to bug reporting
+      console.error('Something went wrong', error);
+    }
+  },
+  /**
+   * Enable query batching
+   */
+  batching: {
+    enabled: true
+  }
+  /**
+   * @link https://trpc.io/docs/caching#api-response-caching
+   */
+  // responseMeta() {
+  //   // ...
+  // },
+});
