+++ src/server/routers/_app.ts
@@ -1,0 +1,19 @@
+import { z } from 'zod';
+import { publicProcedure, router } from '../trpc';
+
+export const appRouter = router({
+  hello: publicProcedure
+    .input(
+      z.object({
+        text: z.string().nullish()
+      })
+    )
+    .query(({ input }) => {
+      return {
+        greeting: `hello ${input?.text ?? 'world'}`
+      };
+    })
+});
+
+// export type definition of API
+export type AppRouter = typeof appRouter;
